# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pokétroc",
    "description": "Application qui permet aux utilisateurs de lister leurs cartes pokémon, de les échanger ainsi que de chercher des cartes qu’ils aimeraient acquérir.",
    "version": "1.1.0",
    "contact": { "url": "https://github.com/nicolaswunderle/poketroc" }
  },
  "servers": [{ "url": "https://poketroc.onrender.com/api" }],
...
  "paths": {
...
    "/dresseurs": {
      "post": {
        "summary": "Créer un dresseur",
        "tags": ["dresseur"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DresseurData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inscription réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dresseur"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/DresseurValidationErreur"
          }
        }
      }
    },
    "/dresseurs?localisation={coordonnéesGeographique}&page={number}&pageSize={number}": {
      "get": {
        "summary": "Afficher les dresseurs par ordre de proximité",
        "tags": ["dresseur"],
        "parameters": [
          {
            "name": "coordonnéesGeographique",
            "in": "path",
            "description": "Coordonnées géographiques du dresseur",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Numéro de la page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Taille de la page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],

        "responses": {
          "200": {
            "description": "Liste des dresseurs proches est affichée.",
            "content": {
              "application/json": {
                "schema": {
                    "$ref": "#/components/schemas/Dresseur"
                  }
                }
            }
          },
          "503": {
            "$ref": "#/components/responses/ErreurConnexion"
          }
        }
      }      
    },
    "dresseurs/{dresseurId}": {
      "get": {
        "summary": "Afficher le dresseur",
        "tags": ["dresseur"],
        "parameters": [
          {
            "name": "dresseurId",
            "in": "path",
            "description": "Identifiant du dresseur",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Le dresseur est affiché",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresseurDataPartiel"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/DresseurEstPasAffiche"
          }
        }
      },
      "patch":{
        "summary": "Modifier partiellement les informations du dresseur",
        "in":"path",
        "tags": ["dresseur"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dresseurId",
            "in": "path",
            "description": "Identifiant du dresseur",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Le dresseur a été modifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresseurDataPartiel"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ModificationErreur"
          }
        }
      },
      "delete": {
        "summary": "Supprimer le dresseur",
        "tags": ["dresseur"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dresseurId",
            "in": "path",
            "description": "Identifiant du dresseur",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Le dresseur a été supprimé"
          },
          "404": {
            "$ref": "#/components/responses/DresseurIntrouvable"
          }
        }
      }
    },
    "/dresseurs/connexion": {
      "post": {
        "summary": "Connexion au compte dresseur",
        "tags": ["dresseur"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/connexion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/connexionReussie"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/connexionEchouee"
          },
          "422": {
            "$ref": "#/components/responses/connexionErreur"
          }
        }
      },

      "delete": {
        "summary": "Deconnexion au compte dresseur",
        "tags": ["dresseur"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "204": {
            "description": "Déconnexion réussie"
          },
          "404": {
            "$ref": "#/components/responses/deconnexionEchouee"
          }
        }
      }
    },
...
    "messages/{echangeId}": {
      "get": {
        "summary": "Afficher tous les messages d'un échange-----------------------------------------------------------",
        "tags": ["message"],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "Conversation d'un échange.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message trouvé.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Message non trouvé."
          }
        }
      },
      "/messages/{messageId}": {
        "delete": {
          "summary": "Supprimer un message",
          "tags": ["message"],
          "parameters": [
            {
              "name": "messageId",
              "in": "path",
              "description": "Identifiant du message",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Le message à été supprimé",
              "content": {
                "text/plain": {
                  "schema": {
                    "example": "Le message à été supprimé"
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/MessageIntrouvable"
            }
          }
        }
      },
      "messages": {
        "post": {
          "summary": "Créer un message",
          "tags": ["message"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Message créé avec succès.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/MessageErreur"
            },
            "422": {
              "$ref": "#/components/responses/DestinataireErreur"
            }
          }
        }
      }
    },
...
    "/messages/{echangeId}":{
      "get": {
          "summary": "Afficger une discussion",
          "tags": ["message"],
          "parameters": [
          {
            "name": "echangeId",
            "in": "path",
            "description": "Identifiant de la carte",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discution affiché",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/DiscutionIntrouvable"
          }
        }
      }
    },
    "/messages/{messageId}": {
      "delete": {
        "summary": "Supprimer un message",
        "tags": ["message"],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "Identifiant du message",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Le message à été supprimé",
            "content": {
              "text/plain": {
                "schema": {
                  "example": "Le message à été supprimé"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/MessageIntrouvable"
          }
        }
      }
    },
    "messages": {
      "post": {
        "summary": "Créer un message",
        "tags": ["message"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message créé avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/MessageErreur"
          },
          "422": {
            "$ref": "#/components/responses/DestinataireErreur"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "DresseurValidationErreur": {
        "description": "Echec de l'inscription.",
        "content": {
          "text/plain": {
            "example": "Inscription échouée."
          }
        }
      },
...
      "connexionReussie": {
        "type": "object",
        "example": "Connexion réussie"
      },
      "connexionErreur": {
        "description": "Mots de passe ou pseudo incorrect",
        "content": {
          "text/plain": {
            "example": "Connexion échouée, veuillez retenter."
          }
        }
      },
      "connexionEchouee": {
        "description": "Mots de passe ou pseudo incorrect",
        "content": {
          "text/plain": {
            "example": "Mots de passe ou pseudo incorrect."
          }
        }
      },
...
      "deconnexionEchouee": {
        "description": "Déconnexion échouée.",
        "content": {
          "text/plain": {
            "example": "Échec de la déconnexion. Une erreur interne s'est produite. Veuillez contacter le support technique pour plus d'assistance."
          }
        }
      },
      "DresseurEstPasAffiche": {
        "description": "Le dresseur n'a pas été affiché.",
        "content": {
          "text/plain": {
            "example": "Le dresseur ne s'affiche pas."
          }
        }
      },
      "ErreurConnexion": {
        "description": "Erreur de connexion.",
        "content": {
          "text/plain": {
            "example": " Le serveur n'est pas disponible en raison d'une surcharge temporaire ou de travaux de maintenance."
          }
        }
      },
      "ModificationErreur": {
        "description": "La modification n'a pas été validé",
        "content": {
          "text/plain": {
            "example": "La modification n'a pas été effectué"
          }
        }
      },
      "DresseurIntrouvable": {
        "description": "Le dresseur est introuvable",
        "content": {
          "text/plain": {
            "example": "Ce dresseur est introuvable"
          }
        }
      },
...
      "DestinataireErreur": {
        "description": "Destinataire introuvable.",
        "content": {
          "text/plain": {
            "example": "Le destinataire n'a pas été trouvé"
          }
        }
      },
      "MessageErreur": {
        "description": "Le message n'a pas été créée",
        "content": {
          "text/plain": {
            "example": "Le message n'a pas pu être créée"
          }
        }
      },
      "MessageIntrouvable": {
        "description": "Message introuvable",
        "content": {
          "text/plain": {
            "example": "Le message est introuvable"
          }
        }
      },
      "DiscutionIntrouvable": {
        "description": "Discution introuvable",
        "content": {
          "text/plain": {
            "example": "La discution est introuvable"
          }
        }
      }
    },
    "schemas": {
...
      "connexion": {
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string",
            "maxlength": 40,
            "unique": true,
            "description": "",
            "example": "Nerak"
          },
          "mot_de_passe": {
            "type": "string",
            "description": "",
            "example": "unMotDePasse88"
          }
        },
        "required": ["pseudo", "mot_de_passe"]
      },
      "Dresseur": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "",
            "example": "7fe91a9a41bc839033eedf0f"
          },
          "nom": {
            "type": "string",
            "description": "",
            "maxlength": 40,
            "example": "Fonseca"
          },
          "prenom": {
            "type": "string",
            "description": "",
            "maxlength": 40,
            "example": "Carina"
          },
          "pseudo": {
            "type": "string",
            "description": "",
            "maxlength": 40,
            "example": "AniraC"
          },
          "email": {
            "type": "string",
            "description": "",
            "example": "adresse@email.ch"
          },
          "age": {
            "type": "integer",
            "description": "",
            "minimum": 1,
            "maximum": 200,
            "example": 25
          },
          "localisation": {
            "type": "Point",
            "description": "",
            "example": "[ -73.856077, 40.848447 ]"
          },
          "image": {
            "type": "string",
            "description": "",
            "example": "https://p1.hiclipart.com/preview/553/126/513/pikachu-i-choose-you-pikachu-icon-png-icon.jpg"
          }
        },
        "required": [
          "id",
          "nom",
          "prenom",
          "pseudo",
          "email",
          "mot_de_passe",
          "age",
          "localisation"
        ]
      },
      "DresseurData": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "description": "",
            "maxlength": 40,
            "example": "Fonseca"
          },
          "prenom": {
            "type": "string",
            "description": "",
            "maxlength": 40,
            "example": "Carina"
          },
          "mot_de_passe": {
            "type": "string",
            "description": "",
            "example": "unMotDePasse32"
          },
          "pseudo": {
            "type": "string",
            "description": "",
            "unique": true,
            "maxlength": 40,
            "example": "AniraC"
          },
          "email": {
            "type": "string",
            "description": "",
            "example": "adresse@email.ch"
          },
          "age": {
            "type": "number",
            "description": "",
            "minimum": 1,
            "maximum": 200,
            "example": 25
          },
          "localisation": {
            "type": "Point",
            "description": "",
            "example": "[ -73.856077, 40.848447 ]"
          }
        },
        "required": ["prenom", "nom", "pseudo", "email", "age", "localisation"]
      },
      "DresseurDataPartiel": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "description": "",
            "minlength": 3,
            "maxlength": 40,
            "example": "Fonseca"
          },
          "prenom": {
            "type": "string",
            "description": "",
            "minlength": 3,
            "maxlength": 15,
            "example": "Carina"
          },
          "pseudo": {
            "type": "string",
            "description": "",
            "example": "AniraC"
          },
          "email": {
            "type": "string",
            "description": "",
            "example": "adresse@email.ch"
          },
          "mot_de_passe": {
            "type": "string",
            "description": "",
            "example": "unMotDePasse32"
          },
          "age": {
            "type": "number",
            "description": "",
            "minimum": 1,
            "maximum": 200,
            "example": 25
          },
          "localisation": {
            "type": "Point",
            "description": "",
            "example": "[ -73.856077, 40.848447 ]"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "",
            "example": "7fe91a9a41bc839033eedf0f"
          },
          "date": {
            "type": "string",
            "description": "",
            "format": "data-time",
            "example": "2020-01-02T10:00:00.000Z"
          },
          "contenu": {
            "type": "string",
            "description": "",
            "example": "Salut!"
          },
          "expéditeur": {
            "type": "string",
            "description": "Celui qui envoie le message.",
            "example": "Nerak"
          },
          "destinataire": {
            "type": "string",
            "description": "Celui qui reçoit le message.",
            "example": "Salocin"
          }
        },
        "required": ["id", "date", "contenu", "destinataire", "expéditeur"]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}